Refactor: Set project properties in dockwidget module
