+++ beoordelingstool_dockwidget.py
@@ -20,16 +20,27 @@
  *                                                                         *
  ***************************************************************************/
 """
-
+import json
 import os
 
 from PyQt4 import QtGui, uic
 from PyQt4.QtCore import pyqtSignal
+from PyQt4.QtGui import QDesktopServices
 from qgis.core import QgsMapLayerRegistry
+from qgis.gui import QgsMessageBar
 from qgis.utils import iface
+
+# Import functions
+from .utils.layer import get_layer_dir
 
 # import constants
 from .utils.constants import HERSTELMAATREGELEN
+# json properties
+from .utils.constants import JSON_NAME
+from .utils.constants import JSON_KEY_PROJ
+from .utils.constants import JSON_KEY_NAME
+from .utils.constants import JSON_KEY_URL
+# Shapefile names
 from .utils.constants import SHP_NAME_MANHOLES
 from .utils.constants import SHP_NAME_PIPES
 from .utils.constants import SHP_NAME_MEASURING_POINTS
@@ -53,10 +64,18 @@ class BeoordelingstoolDockWidget(QtGui.QDockWidget, FORM_CLASS):
         self.setupUi(self)
         self.add_herstelmaatregelen()
         self.tabWidget.currentChanged.connect(self.tab_changed)
+        # General tab
+        self.set_project_properties()
 
     def closeEvent(self, event):
         self.closingPlugin.emit()
         event.accept()
+
+    def add_herstelmaatregelen(self):
+        """Add the herstelmaatregelen to the comboboxes"""
+        self.field_combobox_manholes.addItems(HERSTELMAATREGELEN)
+        self.field_combobox_pipes.addItems(HERSTELMAATREGELEN)
+        self.field_combobox_measuring_points.addItems(HERSTELMAATREGELEN)
 
     def tab_changed(self):
         """Change the active layer upon selecting another tab."""
@@ -73,9 +92,30 @@ class BeoordelingstoolDockWidget(QtGui.QDockWidget, FORM_CLASS):
             measuring_points_layer = QgsMapLayerRegistry.instance().mapLayersByName(SHP_NAME_MEASURING_POINTS)[0]
             iface.setActiveLayer(measuring_points_layer)
 
-    def add_herstelmaatregelen(self):
-        """Add the herstelmaatregelen to the comboboxes"""
-        self.field_combobox_manholes.addItems(HERSTELMAATREGELEN)
-        self.field_combobox_pipes.addItems(HERSTELMAATREGELEN)
-        self.field_combobox_measuring_points.addItems(HERSTELMAATREGELEN)
-
+    def set_project_properties(self):
+        """
+        Set the project name on the General tab of the dockwidget.
+        The name of the project name property of the review.json in the same
+        folder as the layer is used as the project name.
+        """
+        # Check if the manholes, pipes and measuring_points layers exist
+        manholes_layerList = QgsMapLayerRegistry.instance().mapLayersByName(SHP_NAME_MANHOLES)
+        pipes_layerList = QgsMapLayerRegistry.instance().mapLayersByName(SHP_NAME_PIPES)
+        measuring_points_layerList = QgsMapLayerRegistry.instance().mapLayersByName(SHP_NAME_MEASURING_POINTS)
+        if manholes_layerList and pipes_layerList and measuring_points_layerList:
+            # Get project name from the json saved in the same folder as the "manholes" layer
+            layer_dir = get_layer_dir(manholes_layerList[0])
+            json_path = os.path.join(layer_dir, JSON_NAME)
+            try:
+                data = json.load(open(json_path))
+                if data[JSON_KEY_PROJ][JSON_KEY_NAME]:
+                    self.label_project_name.setText(data[JSON_KEY_PROJ][JSON_KEY_NAME])
+                else:
+                    iface.messageBar().pushMessage("Warning", "No project name defined.", level=QgsMessageBar.WARNING, duration=0)
+                if data[JSON_KEY_PROJ][JSON_KEY_URL]:
+                    self.textedit_project_url.setText("<a href=google.com>{}</a>".format(data[JSON_KEY_PROJ][JSON_KEY_URL])).clicked(QDesktopServices.openUrl(QUrl(data[JSON_KEY_PROJ][JSON_KEY_URL], QUrl.TolerantMode)))
+                    # self.label_project_url.setText("<a href={}>{}</a>".format(data[JSON_KEY_PROJ][JSON_KEY_URL]))
+                else:
+                    iface.messageBar().pushMessage("Warning", "No project url defined.", level=QgsMessageBar.WARNING, duration=0)
+            except:
+                iface.messageBar().pushMessage("Error", "No {} found.".format(JSON_NAME), level=QgsMessageBar.CRITICAL, duration=0)
